smprpg/gui: move reforge gui
