+++ SMPRPG/src/main/java/xyz/devvydont/smprpg/commands/player/CommandReforge.java
@@ -8,7 +8,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import xyz.devvydont.smprpg.SMPRPG;
 import xyz.devvydont.smprpg.commands.CommandBase;
-import xyz.devvydont.smprpg.gui.MenuReforge;
+import xyz.devvydont.smprpg.gui.items.MenuReforge;
 import xyz.devvydont.smprpg.reforge.ReforgeBase;
 import xyz.devvydont.smprpg.reforge.ReforgeType;
 import xyz.devvydont.smprpg.util.formatting.ComponentUtils;
